refactor: refactoring the component
