+++ .idea/workspace.xml
@@ -1,0 +1,98 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="AutoImportSettings">
+    <option name="autoReloadType" value="SELECTIVE" />
+  </component>
+  <component name="ChangeListManager">
+    <list default="true" id="692c07b9-abda-40a2-bcd0-e08ea647851c" name="Changes" comment="">
+      <change beforePath="$PROJECT_DIR$/src/components/CharacterLimitMainText/CharacterLimitMainText.module.scss" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/components/CharacterLimitMainText/CharacterLimitMainText.spec.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/components/CharacterLimitMainText/CharacterLimitMainText.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/components/CharacterLimitMainText/CharacterLimitMainText.types.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/components/CharacterLimitMainText/CharacterLimit.types.ts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/components/CharacterLimitMainText/components/CharacterLimit.module.scss" beforeDir="false" afterPath="$PROJECT_DIR$/src/components/CharacterLimitMainText/CharacterLimit.module.scss" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/components/CharacterLimitMainText/components/CharacterLimit.spec.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/components/CharacterLimitMainText/CharacterLimit.spec.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/components/CharacterLimitMainText/components/CharacterLimit.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/components/CharacterLimitMainText/CharacterLimit.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/components/ComposerEditor/ComposerEditor.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/components/ComposerEditor/ComposerEditor.tsx" afterDir="false" />
+    </list>
+    <option name="SHOW_DIALOG" value="false" />
+    <option name="HIGHLIGHT_CONFLICTS" value="true" />
+    <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
+    <option name="LAST_RESOLUTION" value="IGNORE" />
+  </component>
+  <component name="Git.Settings">
+    <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
+  </component>
+  <component name="ProjectColorInfo">{
+  &quot;customColor&quot;: &quot;&quot;,
+  &quot;associatedIndex&quot;: 7
+}</component>
+  <component name="ProjectId" id="2f8cJoPPFxFlMG34AAhcH4GKSog" />
+  <component name="ProjectViewState">
+    <option name="hideEmptyMiddlePackages" value="true" />
+    <option name="showLibraryContents" value="true" />
+  </component>
+  <component name="PropertiesComponent">{
+  &quot;keyToString&quot;: {
+    &quot;RunOnceActivity.ShowReadmeOnStart&quot;: &quot;true&quot;,
+    &quot;Vitest.CharacterLimit.spec.tsx.executor&quot;: &quot;Run&quot;,
+    &quot;git-widget-placeholder&quot;: &quot;issue-382&quot;,
+    &quot;node.js.detected.package.eslint&quot;: &quot;true&quot;,
+    &quot;node.js.detected.package.standard&quot;: &quot;true&quot;,
+    &quot;node.js.detected.package.tslint&quot;: &quot;true&quot;,
+    &quot;node.js.selected.package.eslint&quot;: &quot;(autodetect)&quot;,
+    &quot;node.js.selected.package.standard&quot;: &quot;&quot;,
+    &quot;node.js.selected.package.tslint&quot;: &quot;(autodetect)&quot;,
+    &quot;nodejs_interpreter_path&quot;: &quot;/home/zoldyck/.nvm/versions/node/v18.18.2/bin/node&quot;,
+    &quot;nodejs_package_manager_path&quot;: &quot;pnpm&quot;,
+    &quot;prettierjs.PrettierConfiguration.Package&quot;: &quot;/home/zoldyck/developer/devhat/octopost/node_modules/prettier&quot;,
+    &quot;settings.editor.selected.configurable&quot;: &quot;preferences.general&quot;,
+    &quot;ts.external.directory.path&quot;: &quot;/home/zoldyck/developer/devhat/octopost/node_modules/typescript/lib&quot;,
+    &quot;vue.rearranger.settings.migration&quot;: &quot;true&quot;
+  }
+}</component>
+  <component name="RecentsManager">
+    <key name="MoveFile.RECENT_KEYS">
+      <recent name="$PROJECT_DIR$/src/components/CharacterLimitMainText" />
+    </key>
+  </component>
+  <component name="SharedIndexes">
+    <attachedChunks>
+      <set>
+        <option value="bundled-js-predefined-1d06a55b98c1-74d2a5396914-JavaScript-WS-241.14494.235" />
+      </set>
+    </attachedChunks>
+  </component>
+  <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
+  <component name="TaskManager">
+    <task active="true" id="Default" summary="Default task">
+      <changelist id="692c07b9-abda-40a2-bcd0-e08ea647851c" name="Changes" comment="" />
+      <created>1713186075920</created>
+      <option name="number" value="Default" />
+      <option name="presentableId" value="Default" />
+      <updated>1713186075920</updated>
+      <workItem from="1713186077079" duration="2167000" />
+      <workItem from="1713188258104" duration="20725000" />
+      <workItem from="1713214208764" duration="3040000" />
+    </task>
+    <servers />
+  </component>
+  <component name="TypeScriptGeneratedFilesManager">
+    <option name="version" value="3" />
+  </component>
+  <component name="XDebuggerManager">
+    <breakpoint-manager>
+      <breakpoints>
+        <line-breakpoint enabled="true" type="javascript">
+          <url>file://$PROJECT_DIR$/src/components/ComposerEditor/ComposerEditor.tsx</url>
+          <option name="timeStamp" value="1" />
+        </line-breakpoint>
+        <line-breakpoint enabled="true" type="javascript">
+          <url>file://$PROJECT_DIR$/src/components/CharacterLimitMainText/CharacterLimit.spec.tsx</url>
+          <line>23</line>
+          <option name="temporary" value="true" />
+          <option name="timeStamp" value="2" />
+        </line-breakpoint>
+      </breakpoints>
+    </breakpoint-manager>
+  </component>
+</project>
