+++ src/components/CharacterLimitMainText/CharacterLimit.module.scss
@@ -1,4 +1,4 @@
-@use '../../../styles/global.scss';
+@use '../../styles/global';
 
 .specificFill {
   fill: global.$primaryWhite !important;
