+++ src/components/CharacterLimitMainText/CharacterLimit.types.ts
@@ -1,9 +1,5 @@
 import { ReactNode } from 'react';
 
-export type CharacterLimitMainTextProps = {
-  module: ModuleProps[];
-};
-
 export type ModuleProps = {
   id: string;
   maxLength: number;
