--- src/components/CharacterLimitMainText/CharacterLimitMainText.module.scss
@@ -1,8 +1,0 @@
-.compost {
-  display: flex;
-  gap: 2.8rem;
-
-  justify-content: flex-start;
-
-  padding: 1.7rem;
-}
